good

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,16 +7,16 @@
  */
 int main(void)
 {
-	unsigned int i, j, k, j1, j2, k1, k2;
+	unsigned long i, j, k, j1, j2, k1, k2;
 
 	j = 1;
 	k = 2;
 
-	printf("%d", j);
+	printf("%li", j);
 
 	for (i = 1; i < 91; i++)
 	{
-		printf(", %d", k);
+		printf(", %li", k);
 		k = k + j;
 		j = k - j;
 	}
@@ -27,8 +27,8 @@ int main(void)
 	k2 = k % 1000000000;
 
 	for (i = 92; i < 99; ++i)
-	{	printf("%d", k1 + (k2 / 1000000000));
-		printf("%d", k2 % 1000000000);
+	{	printf("%li", k1 + (k2 / 1000000000));
+		printf("%li", k2 % 1000000000);
 		k1 = k1 + j1;
 		j1 = k1 - j1;
 		k2 = k2 + j2;
